Issue #26: Integer decoding returns negative number
https://github.com/alobbs/libhpack/issues/26

With a big enough number you can make *integer_decode* function return
a negative number. Something that should never happen.

For a 32 bit machine the VLQ encoding **ffff ffff ff0a** gives a
negative number with current code.

We were lucky this didn't affect the hpack library.
The reasons why it didn't matter to the library are:
- For indexed header fields function *hpack_header_table_get* has n
defined as uint32_t and therefore fixes the sign issue.
- For String literals function *parse_string* evaluates is saved because
chula's length is defined as uint32_t in its structure, and therefore
when evaluating expression *buf->len < n + len* it is done as if they
were all usigned; even when len is a negative signed int.

The new test in *integer_test.c* named *decode_too_big_for_int*
illustrates this fact.

This branch is **NOT** to be merged since this is just an example to
show the problem.

There will be another branch that will deal with this issue.